Robot Framework/requirements.txt
robotframework
robotframework-seleniumlibrary
pandas
pyarrow
selenium
webdriver-manager